Rehome the sweep-loop base into a neutral module
